--- tests/data/state0.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" ?>
-<State openmmVersion="7.3" time="0" type="State" version="1">
-	<PeriodicBoxVectors>
-		<A x="2" y="0" z="0"/>
-		<B x="0" y="2" z="0"/>
-		<C x="0" y="0" z="2"/>
-	</PeriodicBoxVectors>
-	<Parameters/>
-	<Energies KineticEnergy="3.805879737208476" PotentialEnergy="18.08819564431735"/>
-	<Positions>
-		<Position x=".0024000000000000002" y="-.0103" z="-.010100000000000001"/>
-		<Position x=".0027" y="-.1132" z="-.0239"/>
-		<Position x="-.08050000000000002" y=".016300000000000002" z=".0471"/>
-		<Position x="-.0059" y=".038400000000000004" z="-.10189999999999999"/>
-		<Position x=".12470000000000002" y=".037500000000000006" z=".0636"/>
-		<Position x=".0814" y=".08610000000000001" z=".14950000000000002"/>
-		<Position x=".2057" y="-.0772" z=".1289"/>
-		<Position x=".31360000000000005" y="-.0752" z=".10320000000000001"/>
-		<Position x=".199" y="-.0641" z=".23950000000000002"/>
-		<Position x=".1656" y="-.17820000000000003" z=".1063"/>
-		<Position x=".1956" y=".1579" z=".0036"/>
-		<Position x=".12190000000000001" y=".2525" z="-.020100000000000003"/>
-		<Position x=".3289" y=".16310000000000002" z="-.020200000000000003"/>
-		<Position x=".39390000000000003" y=".0868" z="-.0174"/>
-		<Position x=".399" y=".2909" z="-.021500000000000002"/>
-		<Position x=".37420000000000003" y=".34400000000000003" z=".06949999999999999"/>
-		<Position x=".3662" y=".38020000000000004" z="-.1434"/>
-		<Position x=".4192" y=".4778" z="-.1358"/>
-		<Position x=".3956" y=".3311" z="-.23820000000000002"/>
-		<Position x=".2577" y=".40270000000000006" z="-.14670000000000002"/>
-		<Position x=".5487000000000001" y=".2654" z="-.0128"/>
-		<Position x=".5889000000000001" y=".1489" z="-.013700000000000002"/>
-		<Position x=".6275000000000001" y=".3733" z="-.0037"/>
-		<Position x=".5963" y=".4691" z="-.0028000000000000004"/>
-		<Position x=".7707" y=".38020000000000004" z=".0068000000000000005"/>
-		<Position x=".8160000000000001" y=".34180000000000005" z="-.0833"/>
-		<Position x=".8233000000000001" y=".3093" z=".1333"/>
-		<Position x=".9342000000000001" y=".3149" z=".13560000000000003"/>
-		<Position x=".7835000000000001" y=".3593" z=".22400000000000003"/>
-		<Position x=".7923" y=".20299999999999999" z=".1332"/>
-		<Position x=".8018000000000001" y=".5323000000000001" z=".013600000000000001"/>
-		<Position x=".7032" y=".6119" z=".012700000000000001"/>
-		<Position x=".9218999999999999" y=".5692" z=".0188"/>
-	</Positions>
-	<Velocities>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-	</Velocities>
-	<Forces>
-		<Force x="-140.13296278919407" y="168.77997539653472" z="-19.82515395130804"/>
-		<Force x="-2.961089066791157" y="1015.653549909367" z="136.21009707239327"/>
-		<Force x="916.7500062630241" y="-294.15621431358795" z="-632.546445817189"/>
-		<Force x="93.93905509641323" y="-551.1845762885933" z="1038.9885852832208"/>
-		<Force x="-438.24559960937154" y="-958.9916025219677" z="-538.9310071149823"/>
-		<Force x="-139.91301749971652" y="157.03862934148316" z="277.5641617373129"/>
-		<Force x="-19.579350495293568" y="13.537068740662107" z="235.83549115583895"/>
-		<Force x="-535.8683295536529" y="-9.932684514432864" z="127.63499601046207"/>
-		<Force x="43.98419565133895" y="-85.99894970634949" z="-726.0674685131494"/>
-		<Force x="205.00707256605136" y="516.3519782835709" z="115.54014563572962"/>
-		<Force x="1016.7691054945469" y="-183.24559415156054" z="-97.93518360325172"/>
-		<Force x="-189.89107311765073" y="243.74078041967113" z="-61.06402215587956"/>
-		<Force x="-863.3580555653791" y="384.3103030924575" z="135.03088957468108"/>
-		<Force x="171.32020574535608" y="-201.10356459031803" z="7.379947324415349"/>
-		<Force x="-204.53372175818362" y="-12.94957600852851" z="-596.126798412337"/>
-		<Force x="-49.61760148533949" y="106.23768705127138" z="182.06458609539902"/>
-		<Force x="39.90026408517588" y="163.65804529663126" z="21.613494872942407"/>
-		<Force x="-314.4864924953379" y="-579.1298427838675" z="-45.09617628235033"/>
-		<Force x="-131.04753628415276" y="218.85830039292208" z="422.561443528493"/>
-		<Force x="517.2481375113194" y="-107.26343865442114" z="15.731971002648539"/>
-		<Force x="179.66337345851605" y="-20.47318933582926" z="11.949792302623335"/>
-		<Force x="-53.5302560483754" y="155.13121466755555" z="1.198438568247212"/>
-		<Force x="-397.06664912344183" y="-238.48454700545312" z="-36.18683871252415"/>
-		<Force x="-27.37377877033552" y="84.05153866019687" z=".789628233759678"/>
-		<Force x="509.29157218381977" y="757.8454861018282" z="672.5160655045753"/>
-		<Force x="130.13442233185722" y="-110.31262290382595" z="-258.83248238632086"/>
-		<Force x="102.99098092428625" y="5.806735689398749" z="113.0073097013557"/>
-		<Force x="-586.676806157316" y="-29.62479814680763" z="-12.167327810296134"/>
-		<Force x="199.3008896735027" y="-250.37800210238993" z="-454.18569581373555"/>
-		<Force x="137.64478266765286" y="471.9883999216593" z=".44401542796011967"/>
-		<Force x="-513.5623077000402" y="-114.81048107153659" z="-26.828320670865935"/>
-		<Force x="736.2739966151458" y="-594.3956402694272" z="6.720553721639259"/>
-		<Force x="-392.37343274843454" y="-120.55436859631327" z="-16.98869150950759"/>
-	</Forces>
-</State>
