--- tests/data/state1.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" ?>
-<State openmmVersion="7.3" time="0" type="State" version="1">
-	<PeriodicBoxVectors>
-		<A x="2" y="0" z="0"/>
-		<B x="0" y="2" z="0"/>
-		<C x="0" y="0" z="2"/>
-	</PeriodicBoxVectors>
-	<Parameters/>
-	<Energies KineticEnergy="2.0858731354589177" PotentialEnergy="70.43384138579417"/>
-	<Positions>
-		<Position x=".0024000000000000002" y="-.0103" z="-.010100000000000001"/>
-		<Position x=".0027" y="-.1132" z="-.0239"/>
-		<Position x="-.08050000000000002" y=".016300000000000002" z=".0471"/>
-		<Position x="-.0059" y=".038400000000000004" z="-.10189999999999999"/>
-		<Position x=".12470000000000002" y=".037500000000000006" z=".0636"/>
-		<Position x=".0814" y=".08610000000000001" z=".14950000000000002"/>
-		<Position x=".2057" y="-.0772" z=".1289"/>
-		<Position x=".31360000000000005" y="-.0752" z=".10320000000000001"/>
-		<Position x=".199" y="-.0641" z=".23950000000000002"/>
-		<Position x=".1656" y="-.17820000000000003" z=".1063"/>
-		<Position x=".1956" y=".1579" z=".0036"/>
-		<Position x=".12190000000000001" y=".2525" z="-.020100000000000003"/>
-		<Position x=".3289" y=".16310000000000002" z="-.020200000000000003"/>
-		<Position x=".39390000000000003" y=".0868" z="-.0174"/>
-		<Position x=".399" y=".2909" z="-.021500000000000002"/>
-		<Position x=".37420000000000003" y=".34400000000000003" z=".06949999999999999"/>
-		<Position x=".3662" y=".38020000000000004" z="-.1434"/>
-		<Position x=".4192" y=".4778" z="-.1358"/>
-		<Position x=".3956" y=".3311" z="-.23820000000000002"/>
-		<Position x=".2577" y=".40270000000000006" z="-.14670000000000002"/>
-		<Position x=".5487000000000001" y=".2654" z="-.0128"/>
-		<Position x=".5889000000000001" y=".1489" z="-.013700000000000002"/>
-		<Position x=".6275000000000001" y=".3733" z="-.0037"/>
-		<Position x=".5963" y=".4691" z="-.0028000000000000004"/>
-		<Position x=".7707" y=".38020000000000004" z=".0068000000000000005"/>
-		<Position x=".8160000000000001" y=".34180000000000005" z="-.0833"/>
-		<Position x=".8233000000000001" y=".3093" z=".1333"/>
-		<Position x=".9342000000000001" y=".3149" z=".13560000000000003"/>
-		<Position x=".7835000000000001" y=".3593" z=".22400000000000003"/>
-		<Position x=".7923" y=".20299999999999999" z=".1332"/>
-		<Position x=".8018000000000001" y=".5323000000000001" z=".013600000000000001"/>
-		<Position x=".7032" y=".6119" z=".012700000000000001"/>
-		<Position x=".9218999999999999" y=".5692" z=".0188"/>
-	</Positions>
-	<Velocities>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-	</Velocities>
-	<Forces>
-		<Force x="-68.98775181061512" y="255.6144884690978" z="-337.4927244247076"/>
-		<Force x="175.42637910598015" y="-5.343302095579238" z="43.65606517151433"/>
-		<Force x="-58.06575067992329" y="-105.05122967557638" z="-35.30223814677116"/>
-		<Force x="131.33291042987744" y="-43.578219699660195" z="-34.99261934576725"/>
-		<Force x="-121.23151313231006" y="-13.776143224820572" z="1097.5933070461658"/>
-		<Force x="-71.79617563483575" y="-202.93368228745305" z="78.62401110805402"/>
-		<Force x="695.6431574143729" y="-213.55681165897136" z="275.3128320786044"/>
-		<Force x="-66.79190193565293" y="229.24720155641097" z="-262.5817827137797"/>
-		<Force x="-4.0596865820239785" y="1.2604436961717767" z="-.39522344050100067"/>
-		<Force x="-203.80186387888142" y="139.16592551123102" z="-260.3231741190795"/>
-		<Force x="-843.5017262615784" y="-168.1814837761967" z="-552.6949874179005"/>
-		<Force x="400.4447262570034" y="267.12310904355985" z="-179.02659112322283"/>
-		<Force x="879.3894625921813" y="488.4046945615765" z="-312.71091547934503"/>
-		<Force x="-785.1498350808722" y="-664.2925991648133" z="124.71927285051231"/>
-		<Force x="463.1826068577688" y="383.9699713979838" z="724.2158252966105"/>
-		<Force x="-233.0222266536981" y="69.58260956163609" z="-104.10755811796253"/>
-		<Force x="-396.5251628131272" y="104.16151184914328" z="-331.6228221962698"/>
-		<Force x="119.13987800123927" y="-72.06828720955204" z="94.66464441928899"/>
-		<Force x="41.650919413376116" y="-121.55546747406152" z="75.87458316170549"/>
-		<Force x="-30.97187843842248" y="-139.83928372279865" z="64.868159638141"/>
-		<Force x="-417.6325349560737" y="-248.87240437848845" z="-196.59799245404014"/>
-		<Force x="325.1678827768445" y="112.00804260998419" z="25.346581739993084"/>
-		<Force x="1309.9426949806145" y="904.2456970892777" z="282.3376741858264"/>
-		<Force x="-1278.6018434255566" y="-415.4266630925626" z="-105.00354512208466"/>
-		<Force x="-63.32608098202418" y="-108.23346119908999" z="-303.4326191256225"/>
-		<Force x="29.2394614708359" y="-245.63509392278795" z="119.3888480717416"/>
-		<Force x="-7.434987005686474" y="-168.36766974719848" z="65.97504977728998"/>
-		<Force x=".9934112711897461" y="-3.734034460575433" z="-38.80813782422586"/>
-		<Force x="-52.813715891110256" y="11.970338311065957" z="-29.774011113776055"/>
-		<Force x="-35.543388586280045" y="10.404838164819477" z="-41.89250745624959"/>
-		<Force x="203.8496978648847" y="436.4719883087448" z="71.20572823847826"/>
-		<Force x="-131.6935211099575" y="-163.2371417398371" z="-9.661445610228606"/>
-		<Force x="95.5483564224607" y="-309.9478816006802" z="-7.36168755239178"/>
-	</Forces>
-</State>
