--- tests/data/state2.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" ?>
-<State openmmVersion="7.3" time="0" type="State" version="1">
-	<PeriodicBoxVectors>
-		<A x="2" y="0" z="0"/>
-		<B x="0" y="2" z="0"/>
-		<C x="0" y="0" z="2"/>
-	</PeriodicBoxVectors>
-	<Parameters/>
-	<Energies KineticEnergy=".03956696232688616" PotentialEnergy="84.63467693048958"/>
-	<Positions>
-		<Position x=".0024000000000000002" y="-.0103" z="-.010100000000000001"/>
-		<Position x=".0027" y="-.1132" z="-.0239"/>
-		<Position x="-.08050000000000002" y=".016300000000000002" z=".0471"/>
-		<Position x="-.0059" y=".038400000000000004" z="-.10189999999999999"/>
-		<Position x=".12470000000000002" y=".037500000000000006" z=".0636"/>
-		<Position x=".0814" y=".08610000000000001" z=".14950000000000002"/>
-		<Position x=".2057" y="-.0772" z=".1289"/>
-		<Position x=".31360000000000005" y="-.0752" z=".10320000000000001"/>
-		<Position x=".199" y="-.0641" z=".23950000000000002"/>
-		<Position x=".1656" y="-.17820000000000003" z=".1063"/>
-		<Position x=".1956" y=".1579" z=".0036"/>
-		<Position x=".12190000000000001" y=".2525" z="-.020100000000000003"/>
-		<Position x=".3289" y=".16310000000000002" z="-.020200000000000003"/>
-		<Position x=".39390000000000003" y=".0868" z="-.0174"/>
-		<Position x=".399" y=".2909" z="-.021500000000000002"/>
-		<Position x=".37420000000000003" y=".34400000000000003" z=".06949999999999999"/>
-		<Position x=".3662" y=".38020000000000004" z="-.1434"/>
-		<Position x=".4192" y=".4778" z="-.1358"/>
-		<Position x=".3956" y=".3311" z="-.23820000000000002"/>
-		<Position x=".2577" y=".40270000000000006" z="-.14670000000000002"/>
-		<Position x=".5487000000000001" y=".2654" z="-.0128"/>
-		<Position x=".5889000000000001" y=".1489" z="-.013700000000000002"/>
-		<Position x=".6275000000000001" y=".3733" z="-.0037"/>
-		<Position x=".5963" y=".4691" z="-.0028000000000000004"/>
-		<Position x=".7707" y=".38020000000000004" z=".0068000000000000005"/>
-		<Position x=".8160000000000001" y=".34180000000000005" z="-.0833"/>
-		<Position x=".8233000000000001" y=".3093" z=".1333"/>
-		<Position x=".9342000000000001" y=".3149" z=".13560000000000003"/>
-		<Position x=".7835000000000001" y=".3593" z=".22400000000000003"/>
-		<Position x=".7923" y=".20299999999999999" z=".1332"/>
-		<Position x=".8018000000000001" y=".5323000000000001" z=".013600000000000001"/>
-		<Position x=".7032" y=".6119" z=".012700000000000001"/>
-		<Position x=".9218999999999999" y=".5692" z=".0188"/>
-	</Positions>
-	<Velocities>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-	</Velocities>
-	<Forces>
-		<Force x="-13.508028211728886" y="2.3940162955306468" z="20.8629290551978"/>
-		<Force x=".875320496341776" y=".21615356684704978" z="-1.592725063743396"/>
-		<Force x=".11117905068777878" y="-1.8831050497428592" z="1.0368415668737212"/>
-		<Force x="-5.7821748668895605" y="7.694434352178068" z="4.604695036451583"/>
-		<Force x="30.300433246929206" y="-159.93375825525288" z="-370.149204115142"/>
-		<Force x="-11.840360500800598" y="-1.5900426794471723" z="-5.068818806327514"/>
-		<Force x="-18.452350532000896" y="19.594756669939827" z="57.30718197755239"/>
-		<Force x="-1.4676113400085047" y="-4.754995310476553" z="-6.531721953613659"/>
-		<Force x="11.350097334985948" y="7.875695601974321" z="-.24526184666779322"/>
-		<Force x="-11.743194494269645" y="1.0071069491952012" z="16.33558837838485"/>
-		<Force x="4.274446963730635" y="111.25805188781662" z="190.52722941139498"/>
-		<Force x="31.26799424052757" y="72.18551388102702" z="190.89866825393557"/>
-		<Force x="28.433297783879233" y="-132.6415631893996" z="-62.426292476511975"/>
-		<Force x="36.11853160181186" y="36.750541953574476" z="162.98635604998628"/>
-		<Force x="-155.53727413804367" y="89.61775300550819" z="-108.95807006525894"/>
-		<Force x="-3.659138926002634" y="20.810849923470638" z="-13.14068984946326"/>
-		<Force x="78.00984258519384" y="-59.684514160539415" z="-64.71328918236695"/>
-		<Force x="4.482282634167152" y="-2.213825885847029" z="-2.8278385726564594"/>
-		<Force x="-2.8204800427488292" y="-1.3063739890879487" z="-.19809230371937936"/>
-		<Force x=".0013934824016585412" y=".0074669643438234345" z=".005095108229114409"/>
-		<Force x="11.874237017430147" y=".10892626613508757" z="-177.7095616242302"/>
-		<Force x="2.7246608946687516" y="1.2592992841683688" z="-41.30777626659006"/>
-		<Force x="-28.74674814033153" y="44.81123872139031" z="363.665877335155"/>
-		<Force x="-.2646160049154811" y="-.12122460734636614" z="3.7303311449098513"/>
-		<Force x="27.798972616511413" y="-163.17459903362814" z="-215.27407374712521"/>
-		<Force x="-13.383853508097728" y="-12.460335018328443" z="-1.4185538203442303"/>
-		<Force x="-9.739456372421945" y="127.07377308956444" z="75.27135112442306"/>
-		<Force x="-.6701386371200748" y="10.694804235421948" z="6.272813538371006"/>
-		<Force x="-11.636834764548032" y="-8.945478664752049" z="-.1749954839185218"/>
-		<Force x="10.67724240487878" y="-3.1079619055452987" z="-6.181639917784873"/>
-		<Force x="11.846241383094277" y="-.6747386234291342" z="-42.39082360855109"/>
-		<Force x="-.4493821750533024" y="-.4066681043361396" z="13.264779283443287"/>
-		<Force x="-.4445310822586864" y="-.4611981709270613" z="13.539691439707061"/>
-	</Forces>
-</State>
