--- tests/data/state3.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" ?>
-<State openmmVersion="7.3" time="0" type="State" version="1">
-	<PeriodicBoxVectors>
-		<A x="2" y="0" z="0"/>
-		<B x="0" y="2" z="0"/>
-		<C x="0" y="0" z="2"/>
-	</PeriodicBoxVectors>
-	<Parameters/>
-	<Energies KineticEnergy=".5793939940598803" PotentialEnergy="290.32038095696885"/>
-	<Positions>
-		<Position x=".0024000000000000002" y="-.0103" z="-.010100000000000001"/>
-		<Position x=".0027" y="-.1132" z="-.0239"/>
-		<Position x="-.08050000000000002" y=".016300000000000002" z=".0471"/>
-		<Position x="-.0059" y=".038400000000000004" z="-.10189999999999999"/>
-		<Position x=".12470000000000002" y=".037500000000000006" z=".0636"/>
-		<Position x=".0814" y=".08610000000000001" z=".14950000000000002"/>
-		<Position x=".2057" y="-.0772" z=".1289"/>
-		<Position x=".31360000000000005" y="-.0752" z=".10320000000000001"/>
-		<Position x=".199" y="-.0641" z=".23950000000000002"/>
-		<Position x=".1656" y="-.17820000000000003" z=".1063"/>
-		<Position x=".1956" y=".1579" z=".0036"/>
-		<Position x=".12190000000000001" y=".2525" z="-.020100000000000003"/>
-		<Position x=".3289" y=".16310000000000002" z="-.020200000000000003"/>
-		<Position x=".39390000000000003" y=".0868" z="-.0174"/>
-		<Position x=".399" y=".2909" z="-.021500000000000002"/>
-		<Position x=".37420000000000003" y=".34400000000000003" z=".06949999999999999"/>
-		<Position x=".3662" y=".38020000000000004" z="-.1434"/>
-		<Position x=".4192" y=".4778" z="-.1358"/>
-		<Position x=".3956" y=".3311" z="-.23820000000000002"/>
-		<Position x=".2577" y=".40270000000000006" z="-.14670000000000002"/>
-		<Position x=".5487000000000001" y=".2654" z="-.0128"/>
-		<Position x=".5889000000000001" y=".1489" z="-.013700000000000002"/>
-		<Position x=".6275000000000001" y=".3733" z="-.0037"/>
-		<Position x=".5963" y=".4691" z="-.0028000000000000004"/>
-		<Position x=".7707" y=".38020000000000004" z=".0068000000000000005"/>
-		<Position x=".8160000000000001" y=".34180000000000005" z="-.0833"/>
-		<Position x=".8233000000000001" y=".3093" z=".1333"/>
-		<Position x=".9342000000000001" y=".3149" z=".13560000000000003"/>
-		<Position x=".7835000000000001" y=".3593" z=".22400000000000003"/>
-		<Position x=".7923" y=".20299999999999999" z=".1332"/>
-		<Position x=".8018000000000001" y=".5323000000000001" z=".013600000000000001"/>
-		<Position x=".7032" y=".6119" z=".012700000000000001"/>
-		<Position x=".9218999999999999" y=".5692" z=".0188"/>
-	</Positions>
-	<Velocities>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-	</Velocities>
-	<Forces>
-		<Force x="-29.22665633776419" y="66.35691755284884" z="-39.08844804157207"/>
-		<Force x="-33.061221665189066" y="1.4333891926107043" z="-32.322468761583586"/>
-		<Force x="-24.825801288843067" y="58.00561497894839" z="-45.208725699205516"/>
-		<Force x="-27.425725726743593" y="103.39182614137837" z="-34.36285115556959"/>
-		<Force x="-14.304649285989726" y="-2.9591124974174514" z="6.810603757843788"/>
-		<Force x="75.5433123868321" y="130.66639577873636" z="-13.211336513286586"/>
-		<Force x="-60.69662769297257" y="-78.46450113761641" z="20.447062422655705"/>
-		<Force x="34.70102382438844" y="-97.46982536065023" z="56.68330478638741"/>
-		<Force x="30.832549514951623" y="-15.033866201814709" z="28.124302366296483"/>
-		<Force x="67.18658410901804" y="-62.274519629289436" z="46.307378336348236"/>
-		<Force x="438.10330243583167" y="209.87135630481274" z="52.91842357561464"/>
-		<Force x="-547.945906766553" y="-792.8295034727608" z="224.0980012875197"/>
-		<Force x="-1242.7811612777073" y="-12.476199669703519" z="-108.91230506365478"/>
-		<Force x="376.14053256022623" y="111.05421425641345" z="-28.109495959805436"/>
-		<Force x="183.99329674967686" y="33.76445281698064" z="2.1469787087189385"/>
-		<Force x="101.46703863649654" y="77.67935754037354" z="33.759337814525615"/>
-		<Force x="-125.3807454817161" y="3.7374369977641795" z="-161.38121271646406"/>
-		<Force x="-52.22706251821856" y="88.24530173951372" z="-63.67763648685661"/>
-		<Force x="11.328019366877044" y="35.165118072523285" z="-32.112319940659155"/>
-		<Force x="144.81086092255364" y="214.89841273639067" z="-154.926376102815"/>
-		<Force x="529.6362324387501" y="652.6464413998519" z="76.02359051197854"/>
-		<Force x="-144.3727101135011" y="-424.5088041460824" z="-34.53717707486358"/>
-		<Force x="-494.43122269380865" y="-1629.009467226078" z="40.80442355615989"/>
-		<Force x="191.73221809226538" y="405.6029593411606" z="28.026153777886023"/>
-		<Force x="92.67007250663782" y="147.99144680126557" z="16.644502733311683"/>
-		<Force x="84.70428266411461" y="165.88508389141305" z="16.67560333822712"/>
-		<Force x="-64.79793154794581" y="-281.28589105398316" z="176.33425820751103"/>
-		<Force x="19.91845230223347" y="28.833304702357665" z="3.1642182690071294"/>
-		<Force x="40.04900256391363" y="18.93336630459018" z="23.63757820943161"/>
-		<Force x="110.75566267576534" y="8.04429398036688" z="59.80151325166428"/>
-		<Force x="646.9575915525547" y="438.9964964889652" z="-119.94291109035703"/>
-		<Force x="-185.7854309057931" y="399.99497580028526" z="3.867608327575219"/>
-		<Force x="-133.26718200034173" y="-4.886472424155272" z="-48.48157863196989"/>
-	</Forces>
-</State>
