--- tests/data/state4.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" ?>
-<State openmmVersion="7.3" time="0" type="State" version="1">
-	<PeriodicBoxVectors>
-		<A x="2" y="0" z="0"/>
-		<B x="0" y="2" z="0"/>
-		<C x="0" y="0" z="2"/>
-	</PeriodicBoxVectors>
-	<Parameters/>
-	<Energies KineticEnergy=".2658782033005881" PotentialEnergy="-728.3355326109192"/>
-	<Positions>
-		<Position x=".0024000000000000002" y="-.0103" z="-.010100000000000001"/>
-		<Position x=".0027" y="-.1132" z="-.0239"/>
-		<Position x="-.08050000000000002" y=".016300000000000002" z=".0471"/>
-		<Position x="-.0059" y=".038400000000000004" z="-.10189999999999999"/>
-		<Position x=".12470000000000002" y=".037500000000000006" z=".0636"/>
-		<Position x=".0814" y=".08610000000000001" z=".14950000000000002"/>
-		<Position x=".2057" y="-.0772" z=".1289"/>
-		<Position x=".31360000000000005" y="-.0752" z=".10320000000000001"/>
-		<Position x=".199" y="-.0641" z=".23950000000000002"/>
-		<Position x=".1656" y="-.17820000000000003" z=".1063"/>
-		<Position x=".1956" y=".1579" z=".0036"/>
-		<Position x=".12190000000000001" y=".2525" z="-.020100000000000003"/>
-		<Position x=".3289" y=".16310000000000002" z="-.020200000000000003"/>
-		<Position x=".39390000000000003" y=".0868" z="-.0174"/>
-		<Position x=".399" y=".2909" z="-.021500000000000002"/>
-		<Position x=".37420000000000003" y=".34400000000000003" z=".06949999999999999"/>
-		<Position x=".3662" y=".38020000000000004" z="-.1434"/>
-		<Position x=".4192" y=".4778" z="-.1358"/>
-		<Position x=".3956" y=".3311" z="-.23820000000000002"/>
-		<Position x=".2577" y=".40270000000000006" z="-.14670000000000002"/>
-		<Position x=".5487000000000001" y=".2654" z="-.0128"/>
-		<Position x=".5889000000000001" y=".1489" z="-.013700000000000002"/>
-		<Position x=".6275000000000001" y=".3733" z="-.0037"/>
-		<Position x=".5963" y=".4691" z="-.0028000000000000004"/>
-		<Position x=".7707" y=".38020000000000004" z=".0068000000000000005"/>
-		<Position x=".8160000000000001" y=".34180000000000005" z="-.0833"/>
-		<Position x=".8233000000000001" y=".3093" z=".1333"/>
-		<Position x=".9342000000000001" y=".3149" z=".13560000000000003"/>
-		<Position x=".7835000000000001" y=".3593" z=".22400000000000003"/>
-		<Position x=".7923" y=".20299999999999999" z=".1332"/>
-		<Position x=".8018000000000001" y=".5323000000000001" z=".013600000000000001"/>
-		<Position x=".7032" y=".6119" z=".012700000000000001"/>
-		<Position x=".9218999999999999" y=".5692" z=".0188"/>
-	</Positions>
-	<Velocities>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-		<Velocity x="0" y="0" z="0"/>
-	</Velocities>
-	<Forces>
-		<Force x="-56.94034596796999" y="-67.30805255558232" z="-22.231070963753027"/>
-		<Force x="-89.80841490070975" y="-63.861587434074856" z="-60.98390165117246"/>
-		<Force x="-146.5792789194643" y="-87.97853116107026" z="-16.912493221885303"/>
-		<Force x="-93.62487899635113" y="-111.44030976523355" z="-69.6594165137946"/>
-		<Force x="101.56800716112194" y="21.73267840831132" z="62.19729545437774"/>
-		<Force x="-6.4756553346300905" y="7.207376163229835" z="80.25605236134687"/>
-		<Force x="122.59852693136507" y="-31.89836704881032" z="63.921278129393805"/>
-		<Force x="-7.324414937455503" y="-27.008217665081983" z="10.497007030695944"/>
-		<Force x="2.010832268813703" y="-19.12435252682851" z="11.488071664473356"/>
-		<Force x="23.053934706161407" y="-21.909424772931356" z="19.76885633766565"/>
-		<Force x="-348.53726806774307" y="-462.8797436429077" z="79.37657098901782"/>
-		<Force x="211.95776265897487" y="680.1432251636189" z="-102.81040031572711"/>
-		<Force x="213.3749821353244" y="277.2486184704253" z="-43.86114848145121"/>
-		<Force x="-224.75185501802693" y="-249.9021180666891" z=".3096682800296617"/>
-		<Force x="-11.902245156009723" y="80.86228628705314" z="-4.952051059035711"/>
-		<Force x="-38.83983769542902" y="-4.402544611342371" z="34.843488534677064"/>
-		<Force x="86.52197885651334" y="80.71234003233617" z="-1.5793024752766367"/>
-		<Force x="-21.138745667012696" y="-23.94286238378895" z="1.0363107408202903"/>
-		<Force x="-22.76772175412348" y="-21.0602503554025" z=".2601705087805408"/>
-		<Force x="-1.1102253294796318" y="-13.444894804354721" z="-8.829982709819054"/>
-		<Force x="-354.85905232895124" y="-87.49621128434197" z="4.881988499564219"/>
-		<Force x="479.1967090172316" y="56.297672776150144" z="-33.171355514299655"/>
-		<Force x="210.32288946628947" y="153.60313166134267" z="-1.167904763695005"/>
-		<Force x="-138.29216144997386" y="-292.7076354288445" z="-31.460559263715464"/>
-		<Force x="-17.600393540454306" y="-61.78358061965922" z="4.367405916817874"/>
-		<Force x="-40.58536247469883" y="-100.80786255701625" z="-28.823268001442116"/>
-		<Force x="21.175824157657136" y="83.47558436599542" z="-52.04691650195697"/>
-		<Force x="-14.251047693530726" y="-62.260914933503" z="26.02944094244811"/>
-		<Force x="-14.824925416954713" y="-42.706307379481714" z="22.445574434327625"/>
-		<Force x="8.111493726859358" y="-13.829950203740712" z="28.62734174529317"/>
-		<Force x="-340.24710087766397" y="-715.2596455189827" z="6.485513332587009"/>
-		<Force x="325.9166613350412" y="618.8259756785714" z="30.695346791252444"/>
-		<Force x="184.65132910527888" y="522.904475712635" z="-8.997610256544782"/>
-	</Forces>
-</State>
