+++ tests/test_bonded.py
@@ -1,313 +1,279 @@
+import functools
 import unittest
+import scipy.linalg
+
 import numpy as np
+import jax
+import jax.numpy as jnp
+from jax.config import config; config.update("jax_enable_x64", True)
 import functools
 
-from jax.config import config; config.update("jax_enable_x64", True)
-from jax.test_util import check_grads
-
-from tests.invariances import assert_potential_invariance
+from common import GradientTest
+from timemachine.lib import ops
 from timemachine.potentials import bonded
 
 
-class TestAngles(unittest.TestCase):
+class TestBonded(GradientTest):
 
-    def test_jax_harmonic_angle_4d(self):
-        x0 = np.array([
-            [ 0.0637,   0.0126,   0.2203,  0.5], # C
-            [ 1.0573,  -0.2011,   1.2864,  0.3], # H
-            [ 2.3928,   1.2209,  -0.2230, -0.2], # H
-            [-0.6891,   1.6983,   0.0780,  0.4], # H
-            [-0.6312,  -1.6261,  -0.2601,  0.9]  # H
-        ], dtype=np.float64)
-        num_atoms = x0.shape[0]
-        params = np.array([75, 1.91, 0.45], dtype=np.float64)
+    def test_restraint(self):
 
-        angle_idxs = np.array([[1,0,2],[1,0,3],[1,0,4],[2,0,3],[2,0,4],[3,0,4]])
-        param_idxs = np.array([[0,1],[0,1],[0,2],[0,1],[0,1],[0,2]])
+        B = 8
 
-        # enable cos angles
-        energy_fn = functools.partial(bonded.harmonic_angle,
-            angle_idxs=angle_idxs,
-            param_idxs=param_idxs,
-            cos_angles=True)
+        params = np.random.randn(B, 3)
 
-        box = None
+        N = 10
+        D = 3
 
-        check_grads(energy_fn, (x0, params, None), order=1, eps=1e-5)
-        check_grads(energy_fn, (x0, params, None), order=2, eps=1e-7)
+        b_idxs = []
 
-        # disable cos angles
-        energy_fn = functools.partial(bonded.harmonic_angle,
-            angle_idxs=angle_idxs,
-            param_idxs=param_idxs,
-            cos_angles=False)
+        for _ in range(B):
+            b_idxs.append(np.random.choice(np.arange(N), size=2, replace=False))
 
-        check_grads(energy_fn, (x0, params, None), order=1, eps=1e-5)
-        check_grads(energy_fn, (x0, params, None), order=2, eps=1e-7) 
+        b_idxs = np.array(b_idxs, dtype=np.int32)
 
-    def test_jax_harmonic_angle(self):
-        x0 = np.array([
-            [ 0.0637,   0.0126,   0.2203], # C
-            [ 1.0573,  -0.2011,   1.2864], # H
-            [ 2.3928,   1.2209,  -0.2230], # H
-            [-0.6891,   1.6983,   0.0780], # H
-            [-0.6312,  -1.6261,  -0.2601], # H
-        ], dtype=np.float64)
-        num_atoms = x0.shape[0]
-        params = np.array([75, 1.91, 0.45], dtype=np.float64)
+        lambda_flags = np.random.randint(0, 2, size=(B,))
 
-        angle_idxs = np.array([[1,0,2],[1,0,3],[1,0,4],[2,0,3],[2,0,4],[3,0,4]])
-        param_idxs = np.array([[0,1],[0,1],[0,2],[0,1],[0,1],[0,2]])
+        for precision, rtol in [(np.float32, 2e-5), (np.float64, 1e-9)]:
 
-        # enable cos angles
-        energy_fn = functools.partial(bonded.harmonic_angle,
-            angle_idxs=angle_idxs,
-            param_idxs=param_idxs,
-            cos_angles=True)
+            ref_nrg = jax.partial(
+                bonded.restraint,
+                lamb_flags=lambda_flags,
+                box=None,
+                bond_idxs=b_idxs
+            )
 
-        box = np.array([
-            [2.0, 0.5, 0.6],
-            [0.6, 1.6, 0.3],
-            [0.4, 0.7, 1.1]
-        ], dtype=np.float64)
+            ref_nrg_params = jax.partial(
+                ref_nrg,
+                params=params
+            )
 
-        assert_potential_invariance(energy_fn, x0, params, box)
+            x_primal = self.get_random_coords(N, D)
+            x_tangent = np.random.randn(*x_primal.shape)
+            lamb_tangent = np.random.rand()
 
-        # disable cos angles
-        energy_fn = functools.partial(bonded.harmonic_angle,
-            angle_idxs=angle_idxs,
-            param_idxs=param_idxs,
-            cos_angles=False)
+            for lamb_primal in [0.0, 0.1, 0.5, 0.7, 1.0]:
 
-        assert_potential_invariance(energy_fn, x0, params, box)
+                # we need to clear the du_dp buffer each time, so we need
+                # to instantiate test_nrg inside here
+                test_nrg = ops.Restraint(
+                    np.array(b_idxs, dtype=np.int32),
+                    np.array(params, dtype=np.float64),
+                    np.array(lambda_flags, dtype=np.int32),
+                    precision=precision
+                )
 
+                self.compare_forces(
+                    x_primal,
+                    lamb_primal,
+                    x_tangent,
+                    lamb_tangent,
+                    ref_nrg_params,
+                    test_nrg,
+                    precision,
+                    rtol
+                )
 
-class TestBonded(unittest.TestCase):
+                primals = (x_primal, lamb_primal, params)
+                tangents = (x_tangent, lamb_tangent, np.zeros_like(params))
 
+                grad_fn = jax.grad(ref_nrg, argnums=(0, 1, 2))
+                ref_primals, ref_tangents = jax.jvp(grad_fn, primals, tangents)
 
-    def test_jax_harmonic_bond_4d(self):
-        x0 = np.array([
-            [1.0, 0.2, 3.3, 1.0], # H 
-            [-0.5,-1.1,-0.9, 0.5], # C
-            [3.4, 5.5, 0.2, -0.3], # H 
-        ], dtype=np.float64)
+                ref_du_dp_primals = ref_primals[2]
+                test_du_dp_primals = test_nrg.get_du_dp_primals()
+                np.testing.assert_almost_equal(ref_du_dp_primals, test_du_dp_primals, rtol)
 
-        params = np.array([10.0, 3.0, 5.5], dtype=np.float64)
-
-        param_idxs = np.array([
-            [0,1],
-            [1,2],
-        ], dtype=np.int32)
-
-        bond_idxs = np.array([
-            [0,1],
-            [1,2]
-        ], dtype=np.int32)
-
-        energy_fn = functools.partial(bonded.harmonic_bond, param_idxs=param_idxs, bond_idxs=bond_idxs)
-
-        check_grads(energy_fn, (x0, params, None), order=1, eps=1e-5)
-        check_grads(energy_fn, (x0, params, None), order=2, eps=1e-7)
-
-    def test_jax_harmonic_bond(self):
-        x0 = np.array([
-            [1.0, 0.2, 3.3], # H 
-            [-0.5,-1.1,-0.9], # C
-            [3.4, 5.5, 0.2], # H 
-        ], dtype=np.float64)
-
-        params = np.array([10.0, 3.0, 5.5], dtype=np.float64)
-
-        param_idxs = np.array([
-            [0,1],
-            [1,2],
-        ], dtype=np.int32)
-
-        bond_idxs = np.array([
-            [0,1],
-            [1,2]
-        ], dtype=np.int32)
-
-        energy_fn = functools.partial(bonded.harmonic_bond, param_idxs=param_idxs, bond_idxs=bond_idxs)
-
-        box = np.array([
-            [2.0, 0.5, 0.6],
-            [0.6, 1.6, 0.3],
-            [0.4, 0.7, 1.1]
-        ], dtype=np.float64)
-
-        assert_potential_invariance(energy_fn, x0, params, box)
+                ref_du_dp_tangents = ref_tangents[2]
+                test_du_dp_tangents = test_nrg.get_du_dp_tangents()
+                np.testing.assert_almost_equal(ref_du_dp_tangents, test_du_dp_tangents, rtol)
 
 
-class TestPeriodicTorsion(unittest.TestCase):
+    def test_harmonic_bond(self):
+        np.random.seed(125)
 
-    def setUp(self):
-        self.conformers = np.array([
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.561317027011325 , 0.2066950040043141, 0.3670430960815993],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 0.9399773448903637,-0.6888774474110431, 0.2104211949995816]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815993],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 1.283345455745044 ,-0.0356257425880843,-0.2573923896494185]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.561317027011325 , 0.2066950040043142, 0.3670430960815992],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 1.263820400176392 , 0.7964992122869241, 0.0084568741589791]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815992],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 0.8993534242298198, 1.042445571242743 , 0.7635483993060286]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.5613170270113255, 0.2066950040043142, 0.3670430960815993],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 0.5250337847650304, 0.476091386095139 , 1.3136545198545133]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.5613170270113255, 0.2066950040043141, 0.3670430960815993],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 0.485009232042489 ,-0.3818599172073237, 1.1530102055165103]],
-            ], dtype=np.float64)
+        N = 64
+        B = 35
+        D = 3
 
-        self.nan_conformers = np.array([
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815993],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 1.2278668040866427, 0.8805184219394547, 0.099391329616366 ]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901],
-             [ 0.561317027011325 , 0.206695004004314 , 0.3670430960815994],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652],
-             [ 0.5494071252089705,-0.5626592973923106, 0.9817919758125693]],
-            ], dtype=np.float64)
+        x = self.get_random_coords(N, D)
 
-        self.conformers_4d = np.array([
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.4],
-             [ 0.561317027011325 , 0.2066950040043141, 0.3670430960815993, 0.3],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.1],
-             [ 0.9399773448903637,-0.6888774474110431, 0.2104211949995816, 0.9]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901,-0.5],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815993, 0.4],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.2],
-             [ 1.283345455745044 ,-0.0356257425880843,-0.2573923896494185, 0.3]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.4],
-             [ 0.561317027011325 , 0.2066950040043142, 0.3670430960815992, 0.1],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.7],
-             [ 1.263820400176392 , 0.7964992122869241, 0.0084568741589791, 0.8]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.1],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815992, 0.1],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.4],
-             [ 0.8993534242298198, 1.042445571242743 , 0.7635483993060286,-0.3]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.5],
-             [ 0.5613170270113255, 0.2066950040043142, 0.3670430960815993, 0.6],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.9],
-             [ 0.5250337847650304, 0.476091386095139 , 1.3136545198545133, 0.2]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.0],
-             [ 0.5613170270113255, 0.2066950040043141, 0.3670430960815993, 0.1],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.4],
-             [ 0.485009232042489 ,-0.3818599172073237, 1.1530102055165103, 0.3]],
-            ], dtype=np.float64)
+        atom_idxs = np.arange(N)
+        bond_params = np.random.rand(B, 2).astype(np.float64)
+        bond_idxs = []
+        for _ in range(B):
+            bond_idxs.append(np.random.choice(atom_idxs, size=2, replace=False))
+        bond_idxs = np.array(bond_idxs, dtype=np.int32)
 
-        self.nan_conformers_4d = np.array([
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.5],
-             [ 0.5613170270113252, 0.2066950040043142, 0.3670430960815993, 0.2],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.3],
-             [ 1.2278668040866427, 0.8805184219394547, 0.099391329616366 , 0.6]],
-            [[-0.6000563454193615, 0.376172954382274 ,-0.2487295756125901, 0.1],
-             [ 0.561317027011325 , 0.206695004004314 , 0.3670430960815994,-0.2],
-             [-1.187055522272264 ,-0.3415864358441354, 0.0871382207830652, 0.3],
-             [ 0.5494071252089705,-0.5626592973923106, 0.9817919758125693, 0.9]],
-            ], dtype=np.float64)
+        lamb = 0.0
 
-    def test_jax_torsions_4d(self):
-        """
-        Test agreement of torsions with OpenMM's implementation of torsion terms.
-        """
-        torsion_idxs = np.array([
-            [0, 1, 2, 3],
-            [0, 1, 2, 3],
-            [0, 1, 2, 3],
-        ], dtype=np.int32)
+        for precision, rtol in [(np.float32, 2e-5), (np.float64, 1e-9)]:
 
-        params = np.array([
-            2.3, # k0
-            5.4, # k1
-            9.0, # k2
-            0.0, # t0
-            3.0, # t1
-            5.8, # t2
-            1.0, # n0
-            2.0, # n1
-            3.0  # n2
-        ])
+            custom_bonded = ops.HarmonicBond(
+                bond_idxs,
+                bond_params,
+                precision=precision
+            )
 
-        param_idxs = np.array([
-            [0, 3, 6],
-            [1, 4, 7],
-            [2, 5, 8]
-        ], dtype=np.int32)
+            # test the parameter derivatives for correctness.
+            harmonic_bond_fn = functools.partial(bonded.harmonic_bond, box=None, bond_idxs=bond_idxs)
+            grad_fn = jax.grad(harmonic_bond_fn, argnums=(0, 1, 2))
 
-        box = np.array([
-            [2.0, 0.5, 0.6],
-            [0.6, 1.6, 0.3],
-            [0.4, 0.7, 1.1]
-        ], dtype=np.float64)
+            harmonic_bond_fn_parameterized = functools.partial(
+                harmonic_bond_fn,
+                params=bond_params
+            )
 
-        energy_fn = functools.partial(
-            bonded.periodic_torsion,
-            param_idxs=param_idxs,
-            torsion_idxs=torsion_idxs)
+            x_tangent = np.random.randn(*x.shape)
+            lamb_tangent = np.random.rand()
 
-        # there's no good finite difference tests that we can do for the nan_conformers
-        # so instead we compare against OpenMM implementation later on
-        for conf_idx, conf in enumerate(self.conformers_4d):
-            check_grads(energy_fn, (conf, params, None), order=1, eps=1e-5)
-            check_grads(energy_fn, (conf, params, None), order=2, eps=1e-7)
+            self.compare_forces(
+                x,
+                lamb,
+                x_tangent,
+                lamb_tangent,
+                harmonic_bond_fn_parameterized,
+                custom_bonded,
+                precision,
+                rtol
+            )
 
-    def test_jax_torsions(self):
-        """
-        Test agreement of torsions with OpenMM's implementation of torsion terms.
-        """
-        torsion_idxs = np.array([
-            [0, 1, 2, 3],
-            [0, 1, 2, 3],
-            [0, 1, 2, 3],
-        ], dtype=np.int32)
+            primals = (x, lamb, bond_params)
+            tangents = (x_tangent, lamb_tangent, np.zeros_like(bond_params))
 
-        params = np.array([
-            2.3, # k0
-            5.4, # k1
-            9.0, # k2
-            0.0, # t0
-            3.0, # t1
-            5.8, # t2
-            1.0, # n0
-            2.0, # n1
-            3.0  # n2
-        ])
+            ref_primals, ref_tangents = jax.jvp(grad_fn, primals, tangents)
 
-        param_idxs = np.array([
-            [0, 3, 6],
-            [1, 4, 7],
-            [2, 5, 8]
-        ], dtype=np.int32)
+            ref_du_dp_primals = ref_primals[2]
+            test_du_dp_primals = custom_bonded.get_du_dp_primals()
+            np.testing.assert_almost_equal(ref_du_dp_primals, test_du_dp_primals, rtol)
 
-        box = np.array([
-            [2.0, 0.5, 0.6],
-            [0.6, 1.6, 0.3],
-            [0.4, 0.7, 1.1]
-        ], dtype=np.float64)
+            ref_du_dp_tangents = ref_tangents[2]
+            test_du_dp_tangents = custom_bonded.get_du_dp_tangents()
+            np.testing.assert_almost_equal(ref_du_dp_tangents, test_du_dp_tangents, rtol)
 
-        energy_fn = functools.partial(
-            bonded.periodic_torsion,
-            param_idxs=param_idxs,
-            torsion_idxs=torsion_idxs)
 
-        # there's no good finite difference tests that we can do for the nan_conformers
-        # so instead we compare against OpenMM implementation later on
-        for conf_idx, conf in enumerate(self.conformers):
-            assert_potential_invariance(energy_fn, conf, params, box)
+    def test_harmonic_angle(self):
+        np.random.seed(125)
 
-if __name__ == "__main__":
-    unittest.main()
+        N = 64
+        A = 25
+        D = 3
 
+        x = self.get_random_coords(N, D)
+
+        atom_idxs = np.arange(N)
+        angle_params = np.random.rand(A, 2).astype(np.float64)
+        angle_idxs = []
+        for _ in range(A):
+            angle_idxs.append(np.random.choice(atom_idxs, size=3, replace=False))
+        angle_idxs = np.array(angle_idxs, dtype=np.int32)
+
+        lamb = 0.0
+
+        for precision, rtol in [(np.float32, 2e-5), (np.float64, 1e-9)]:
+
+            custom_angle = ops.HarmonicAngle(
+                angle_idxs,
+                angle_params,
+                precision=precision
+            )
+
+            # test the parameter derivatives for correctness.
+            harmonic_angle_fn = functools.partial(bonded.harmonic_angle, box=None, angle_idxs=angle_idxs)
+            grad_fn = jax.grad(harmonic_angle_fn, argnums=(0, 1, 2))
+
+            harmonic_angle_fn_parameterized = functools.partial(
+                harmonic_angle_fn,
+                params=angle_params
+            )
+
+            x_tangent = np.random.randn(*x.shape)
+            lamb_tangent = np.random.rand()
+
+            self.compare_forces(
+                x,
+                lamb,
+                x_tangent,
+                lamb_tangent,
+                harmonic_angle_fn_parameterized,
+                custom_angle,
+                precision,
+                rtol
+            )
+
+            primals = (x, lamb, angle_params)
+            tangents = (x_tangent, lamb_tangent, np.zeros_like(angle_params))
+
+            ref_primals, ref_tangents = jax.jvp(grad_fn, primals, tangents)
+
+            ref_du_dp_primals = ref_primals[2]
+            test_du_dp_primals = custom_angle.get_du_dp_primals()
+            np.testing.assert_almost_equal(ref_du_dp_primals, test_du_dp_primals, rtol)
+
+            ref_du_dp_tangents = ref_tangents[2]
+            test_du_dp_tangents = custom_angle.get_du_dp_tangents()
+            np.testing.assert_almost_equal(ref_du_dp_tangents, test_du_dp_tangents, rtol)
+
+
+    def test_periodic_torsion(self):
+        np.random.seed(125)
+
+        N = 64
+        T = 25
+        D = 3
+
+        x = self.get_random_coords(N, D)
+
+        atom_idxs = np.arange(N)
+        torsion_params = np.random.rand(T, 3).astype(np.float64)
+        torsion_idxs = []
+        for _ in range(T):
+            torsion_idxs.append(np.random.choice(atom_idxs, size=4, replace=False))
+
+        torsion_idxs = np.array(torsion_idxs, dtype=np.int32)
+
+        lamb = 0.0
+
+        for precision, rtol in [(np.float32, 2e-5), (np.float64, 1e-9)]:
+
+            custom_torsion = ops.PeriodicTorsion(
+                torsion_idxs,
+                torsion_params,
+                precision=precision
+            )
+
+            # test the parameter derivatives for correctness.
+            periodic_torsion_fn = functools.partial(bonded.periodic_torsion, box=None, torsion_idxs=torsion_idxs)
+            grad_fn = jax.grad(periodic_torsion_fn, argnums=(0, 1, 2))
+
+            periodic_torsion_fn_parameterized = functools.partial(
+                periodic_torsion_fn,
+                params=torsion_params
+            )
+
+            x_tangent = np.random.randn(*x.shape)
+            lamb_tangent = np.random.rand()
+
+            self.compare_forces(
+                x,
+                lamb,
+                x_tangent,
+                lamb_tangent,
+                periodic_torsion_fn_parameterized,
+                custom_torsion,
+                precision,
+                rtol
+            )
+
+            primals = (x, lamb, torsion_params)
+            tangents = (x_tangent, lamb_tangent, np.zeros_like(torsion_params))
+
+            ref_primals, ref_tangents = jax.jvp(grad_fn, primals, tangents)
+
+            ref_du_dp_primals = ref_primals[2]
+            test_du_dp_primals = custom_torsion.get_du_dp_primals()
+            np.testing.assert_almost_equal(ref_du_dp_primals, test_du_dp_primals, rtol)
+
+            ref_du_dp_tangents = ref_tangents[2]
+            test_du_dp_tangents = custom_torsion.get_du_dp_tangents()
+            np.testing.assert_almost_equal(ref_du_dp_tangents, test_du_dp_tangents, rtol)
